+++ spec/features/conversions/users_can_complete_conversion_tasks_spec.rb
@@ -30,6 +30,7 @@ RSpec.feature "Users can complete conversion tasks" do
     risk_protection_arrangement
     sponsored_support_grant
     conditions_met
+    confirm_headteacher_contact
     main_contact
     proposed_capacity_of_the_academy
     receive_grant_payment_certificate