+++ spec/models/conversion/task_list_spec.rb
@@ -13,6 +13,7 @@ RSpec.describe Conversion::TaskList do
         :conversion_grant,
         :sponsored_support_grant,
         :academy_details,
+        :confirm_headteacher_contact,
         :main_contact,
         :chair_of_governors_contact,
         :proposed_capacity_of_the_academy,
@@ -57,6 +58,7 @@ RSpec.describe Conversion::TaskList do
               Conversion::Task::ConversionGrantTaskForm,
               Conversion::Task::SponsoredSupportGrantTaskForm,
               Conversion::Task::AcademyDetailsTaskForm,
+              Conversion::Task::ConfirmHeadteacherContactTaskForm,
               Conversion::Task::MainContactTaskForm,
               Conversion::Task::ChairOfGovernorsContactTaskForm,
               Conversion::Task::ProposedCapacityOfTheAcademyTaskForm
@@ -120,7 +122,7 @@ RSpec.describe Conversion::TaskList do
       project = create(:conversion_project)
       task_list = described_class.new(project, user)
 
-      expect(task_list.tasks.count).to eql 31
+      expect(task_list.tasks.count).to eql 32
       expect(task_list.tasks.first).to be_a Conversion::Task::HandoverTaskForm
       expect(task_list.tasks.last).to be_a Conversion::Task::ReceiveGrantPaymentCertificateTaskForm
     end