+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- a new 'confirm the headteacher contact' task has been added, users need to
+  choose the appropriate contact from those available to the project, they may
+  also have to add it
+
 ## [Release-82][release-82]
 
 ### Added